move isUserAuthorised method inside the roles controller
